% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Amberg2011.r
\name{AijklWeights}
\alias{AijklWeights}
\title{subfunction of AmbergDeformSpam}
\usage{
AijklWeights(mesh, arcnode)
}
\arguments{
\item{mesh}{triangular mesh}

\item{arcnode}{output from createArcNode(mesh)}
}
\description{
calculate per face-deform weights for approx. derivation
}

