% Generated by roxygen2 (4.0.1): do not edit by hand
\name{areafun}
\alias{areafun}
\title{subfunction of AmbergDeformSpam}
\usage{
areafun(x)
}
\arguments{
\item{x}{n x 12  matrix containing cornerstones of area}
}
\description{
calculate area between edge and barycenters of adjacent faces needed in AijkWeights
}

