% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Amberg2011.r
\name{buildAffineAmberg}
\alias{buildAffineAmberg}
\title{subfunction of AmbergDeformSpam}
\usage{
buildAffineAmberg(mesh)
}
\arguments{
\item{mesh}{triangular mesh}
}
\description{
subroutine preparing vertex assortment in order to create S using an 'arcface' matrix
}

