% Generated by roxygen2 (4.0.1): do not edit by hand
\name{closeMeshes}
\alias{closeMeshes}
\title{close two meshes where one is only an offset of the other}
\usage{
closeMeshes(mesh1, mesh2, invert = TRUE)
}
\arguments{
\item{mesh1}{mesh}

\item{mesh2}{mesh}

\item{invert}{logical: if TRUE the orientation of mesh1 will be flipped}
}
\description{
close two meshes where one is only an offset of the other
}

