% Generated by roxygen2 (4.0.1): do not edit by hand
\name{createBayes}
\alias{createBayes}
\title{create an object of class BayesDeform to be used in gaussMatch and AmbergRegister}
\usage{
createBayes(model, sdmax = numeric(0), mahalanobis = "chisq",
  ptValueNoise = 2)
}
\arguments{
\item{model}{a model of class pPCA created with package RvtkStatismo}

\item{sdmax}{numeric or vector of numerics (length must be of value =< \code{iterations} as to be specified in gaussMatch or AmbergRegister. If \code{length(sd) =n < iterations}, then the first n iterations will be restricted to the according standard deviations}

\item{mahalanobis}{parameter according to mahaprob in function PredictSample from RvtkStatismo}

\item{ptValueNoise}{parameter according to ptValueNoise in function statismoConstrainModel from RvtkStatismo}
}
\value{
returns a list of class "BayesDeform"
}
\description{
create an object of class BayesDeform to be used in gaussMatch and AmbergRegister
}

