% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Amberg2011.r
\name{createC}
\alias{createC}
\title{subfunction of AmbergDeformSpam}
\usage{
createC(lm1, mesh)
}
\arguments{
\item{lm1}{landmarks}

\item{mesh}{triangular mesh}
}
\description{
create C - middle block of Jacobian matrix
}

