% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displacementField.r
\name{createDisplacementField}
\alias{createDisplacementField}
\title{create a discrete displacement field}
\usage{
createDisplacementField(reference, target)
}
\arguments{
\item{reference}{k x 3 matrix containing coordinates or a mesh of classe "mesh3d"}

\item{target}{k x 3 matrix or a mesh of classe "mesh3d"}
}
\value{
returns an object of class "DisplacementField", which is a list containing
\item{domain}{k x matrix containing starting points of the displacement field}
\item{DisplacementField}{k x matrix containing direction vectors of the displacement field}
}
\description{
create a discrete displacement field based on two sets of coordinates/meshes
}
\examples{
require(Rvcg)
data(dummyhead)
humoff <- meshOffset(dummyhead.mesh,offset=5)
dispfield <- createDisplacementField(dummyhead.mesh,humoff)
}

