% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Amberg2011.r
\name{createJc}
\alias{createJc}
\title{subfunction of AmbergDeformSpam}
\usage{
createJc(lm1, ncol, mesh)
}
\arguments{
\item{lm1}{landmarks}

\item{ncol}{column numbers of complete Jacbian}

\item{mesh}{triangular mesh}
}
\description{
function to setup Jc Jacobian submatrix
}

