% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insideBBox.r
\name{cropOutsideBBox}
\alias{cropOutsideBBox}
\title{remove parts of a mesh outside a bounding box defined by another mesh}
\usage{
cropOutsideBBox(mesh1, mesh2, extend = 0, pca = TRUE, outside = TRUE)
}
\arguments{
\item{mesh1}{reference mesh of which the bounding box will be computed}

\item{mesh2}{mesh to be cropped}

\item{extend}{numeric: offset extending the bbox (or shrinking if negative)}

\item{pca}{if TRUE, the box will be aligned by the principal axes.}

\item{outside}{logical: if TRUE the part outside the box is removed}
}
\value{
returns the cropped mesh2
}
\description{
remove parts of a mesh outside a bounding box defined by another mesh
}

