% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remeshList.r
\name{decimateList}
\alias{decimateList}
\title{decimate a list of registered meshes (with corresponding vertices)}
\usage{
decimateList(matchlist, reference = 1, random = FALSE, ...)
}
\arguments{
\item{matchlist}{a list of meshes with corresponding vertices (same amount and pseudo-homologous positions), e.g. registered with \code{\link{gaussMatch}}. Meshes do not need to be aligned.}

\item{reference}{integer: select the specimen the to which the decimation is to applied initially.}

\item{random}{if TRUE, a random specimen is selected for initial decimation}

\item{...}{parameters passed to vcgQEdecim}
}
\value{
a list of decimated meshes with correspondences preserved.
}
\description{
decimate a list of registered meshes (with corresponding vertices), preserving correspondences using Quadric Edge Collapse decimation
}
\details{
The decimation is applied to a reference and then the barycentric coordinates of the new vertices on the original surface are calculated. These are used to extract the corresponding positions of the remeshed versions on all meshes in the sample. The Decimation is performed by the function \code{vcgQEdecim} from the Rvcg-package.
}
