% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displacementField.r
\name{displacementField2Grid}
\alias{displacementField2Grid}
\title{convert a discrete irregular displacement field to a displacement grid}
\usage{
displacementField2Grid(
  x,
  spacing = rep(2, 3),
  margin = 0.05,
  IJK2RAS = diag(c(-1, -1, 1, 1)),
  invert = FALSE,
  ...
)
}
\arguments{
\item{x}{field of class "DisplacementField", e.g. created using \code{\link{createDisplacementField}}}

\item{spacing}{spacing between grid nodes}

\item{margin}{percentage of offset around \code{x}}

\item{IJK2RAS}{image to coordinate transform}

\item{invert}{if TRUE, the displacement field will be inverted}

\item{...}{additional arguments passed to \code{\link{interpolateDisplacementField}} to choose interpolation etc..}
}
\value{
an object of class DisplacementField with additional attributes
}
\description{
convert a discrete irregular displacement field to a displacement grid
}
