% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getCoordVar}
\alias{getCoordVar}
\alias{getCoordVar.pPCA}
\alias{getCoordVar.pPCAcond}
\title{get per coordinate variance from a statistical model}
\usage{
getCoordVar(model)

\method{getCoordVar}{pPCA}(model)

\method{getCoordVar}{pPCAcond}(model)
}
\arguments{
\item{model}{object of class pPCA}
}
\description{
get per coordinate variance from a statistical model
}

