% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/insideBBox.r
\name{getMeshBox}
\alias{getMeshBox}
\title{compute the bounding box of a triangular mesh}
\usage{
getMeshBox(mesh, tri = FALSE, extend = 0)
}
\arguments{
\item{mesh}{triangular mesh of class "mesh3d"}

\item{tri}{logical: if TRUE, triangular mesh representing the bounding box is returned}

\item{extend}{numeric: offset extending the bbox (or shrinking if negative)}
}
\value{
returns a matrix with the corners of the box or a triangular mesh if tri=TRUE.
}
\description{
compute the bounding box of a triangular mesh
}
\examples{
require(Rvcg)
require(rgl)
data(humface)
bbox <- getMeshBox(humface)
spheres3d(bbox)
bbmesh <- getMeshBox(humface,tri=TRUE)
shade3d(bbmesh,alpha=0.4)
wire3d(humface)
}

