% Generated by roxygen2 (4.0.0): do not edit by hand
\name{getProb}
\alias{getCoefficients}
\alias{getProb}
\alias{getProb.matrix}
\alias{getProb.mesh3d}
\title{calculate probability/coefficients for a matrix/mesh given a statistical model}
\usage{
getProb(x, model, use.lm)

\method{getProb}{matrix}(x, model, use.lm = NULL)

\method{getProb}{mesh3d}(x, model, use.lm = NULL)

getCoefficients(x, model, use.lm = NULL)
}
\arguments{
\item{x}{matrix or mesh3d}

\item{model}{a model of class pPCA}

\item{use.lm}{integer vector specifying row indices of the coordinates to use for rigid registration on the model's meanshape.}
}
\value{
\code{getProb} returns a probability, while \code{getCoefficients} returns the (scaled) scores in the pPCA space.
}
\description{
calculate probability for a matrix/mesh given a statistical model
}

