% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorMesh.r
\name{getVertColor}
\alias{getVertColor}
\title{get the per-vertex colors of an object of class mesh3d}
\usage{
getVertColor(x)
}
\arguments{
\item{x}{mesh3d}
}
\value{
vector of hexadecimal rgb values
}
\description{
get the per-vertex colors of an object of class mesh3d
}

