% Generated by roxygen2 (4.0.1): do not edit by hand
\name{writeh5}
\alias{h5model.read}
\alias{writeh5}
\alias{writeh5.pPCA}
\title{Im- and Export a probabilistic PCA model into the statismo format}
\usage{
writeh5(model, filename)

\method{writeh5}{pPCA}(model, filename)

h5model.read(filename)
}
\arguments{
\item{model}{PCA model}

\item{filename}{character specifying filename}
}
\description{
Im- and Export a probabilistic PCA model into the statismo format
}
\references{
https://docs.google.com/document/d/1ySwSsD-di8ez5WnViLl4VsYeRI-ggVa-HQRxG4kLjH4/edit
}

