\name{mesheR-internal}

\alias{gaussDisplace}
\alias{placeDowels}
\alias{glView}
\alias{glVisible}
\alias{cylinder}
%\alias{plotspin}
%\alias{spinimage}

%curvature
\alias{mesh2ply2}
\alias{plotCurves}
\alias{readCurves}
\alias{file2ply2}
%face helpers
\alias{add.faces}
\alias{del.faces}
\alias{find.faces}



\title{
  low level functions of mesheR}
\description{Internal function(s)}
\keyword{internal}



