% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displacementField.r
\name{invertDisplacementField}
\alias{invertDisplacementField}
\title{invert a displacement field}
\usage{
invertDisplacementField(dispfield)
}
\arguments{
\item{dispfield}{displacement field of class "DisplacementField", e.g. created using \code{\link{createDisplacementField}}}
}
\value{
returns an inverted displacement field
}
\description{
invert a displacement field
}
\examples{
require(Rvcg);require(Morpho)
data(dummyhead)
humoff <- meshOffset(dummyhead.mesh,offset=5)
dispfield <- createDisplacementField(dummyhead.mesh,humoff)
dispfield_inverted <- invertDisplacementField(dispfield)
}

