% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Amberg2011.r
\name{mat2sparseBlock}
\alias{mat2sparseBlock}
\title{subfunction of AmbergDeformSpam}
\usage{
mat2sparseBlock(mat)
}
\arguments{
\item{mat}{quadratic matrix}
}
\description{
convert affine per-face trafo-matrix
}

