% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanMeshes.r
\name{meanMeshes}
\alias{meanMeshes}
\title{average the vertices from a set of corresponding meshes}
\usage{
meanMeshes(...)
}
\arguments{
\item{...}{meshes of class mesh3d}
}
\value{
a mesh with averaged vertices
}
\description{
average the vertices from a set of corresponding meshes
}

