% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesheR-deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{plotDisplacementField}
\title{deprecated functions of mesheR}
\usage{
plotDisplacementField(...)
}
\description{
document deprecated functions
}
\keyword{internal}

