% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelFitting.r
\name{miniSQmodel}
\alias{miniSQmodel}
\title{minimize mean squared distance between model and a point-cloud with correspondences}
\usage{
miniSQmodel(clost, model, iterations = 10, initpar = NULL, use = NULL,
  sdmax = NULL, mahaprob = c("none", "chisq", "dist"), ...)
}
\arguments{
\item{clost}{matrix or mesh3d}

\item{model}{statismo model of class pPCA}

\item{iterations}{integer: max number of iterations passed to lbfgs}

\item{initpar}{initial estimate of the model parameters}

\item{use}{integer vector: which points to use}

\item{sdmax}{constrain parameters (normalized PC-scores) to be within +- sdmax}

\item{mahaprob}{character: if != "none", use mahalanobis-distance to determine overall probability (of the shape projected into the model space."chisq" uses the Chi-Square distribution of the squared Mahalanobisdistance, while "dist" restricts the values to be within a multi-dimensional sphere of radius \code{sdmax}. If FALSE the probability will be determined per PC separately.}

\item{...}{additional parameters to be passed to \code{\link{lbfgs}}.}
}
\value{
\item{par}{the model's parameters}
\item{mesh}{the fitted mesh}
}
\description{
minimize mean squared distance between model and a point-cloud with correspondences
}

