% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixColorMesh.r
\name{mixColorMesh}
\alias{mixColorMesh}
\title{mix vertex colors of two registered meshes.}
\usage{
mixColorMesh(mesh1, mesh2, alpha = 0.5)
}
\arguments{
\item{mesh1}{triangular mesh of class "mesh3d".}

\item{mesh2}{triangular mesh of class "mesh3d".}

\item{alpha}{numeric: 0 <= alpha <=1 weight of color associated with mesh2.}
}
\value{
returns mesh1 with mixed vertex colors.
}
\description{
mixColorMesh mixes the vertex colors of two
registered meshes (identical faces and corresponding vertices).
}
\examples{

require(Morpho)
data(nose)
redmesh <- shortnose.mesh
redmesh$material$color <- rep("#FF0000",,ncol(shortnose.mesh$vb))
bluemesh <- shortnose.mesh
bluemesh$material$color <- rep("#0000FF",ncol(shortnose.mesh$vb))
mixmesh <- mixColorMesh(bluemesh,redmesh)
\dontrun{
require(rgl)
shade3d(mixmesh)
}
}
\author{
Stefan Schlager
}
