% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modelFitting.r
\name{modelFitting}
\alias{modelFitting}
\title{fit a model minimizing the (symmetric) mean squared  distance}
\usage{
modelFitting(model, tarmesh, iterations = 5, symmetric = TRUE,
  tardist = 1e+05, silent = FALSE)
}
\arguments{
\item{model}{statistical model of class pPCA}

\item{tarmesh}{a target mesh already aligned to the model}

\item{iterations}{numbers of iterations to run}

\item{symmetric}{logical: if TRUE, the symmetric distance will be minimized.}

\item{tardist}{maximal distance from target to model to be considered}

\item{silent}{logical: if TRUE output will be suppressed.}
}
\value{
\item{par}{the model's parameters}
\item{mesh}{the fitted mesh}
}
\description{
fit a model minimizing the (symmetric) mean squared distance
}
\details{
this function fits a statistical model to a target mesh using a l-bfgs optimizer to minimize the symmetric mean squared distance between meshes.
}
\note{
needs RvtkStatismo installed
}
\examples{
require(RvtkStatismo)
download.file(url="https://github.com/marcelluethi/statismo-shaperegistration/raw/master/data/VSD001_femur.vtk","./VSD001_femur.vtk",method = "w")
download.file(url="https://github.com/marcelluethi/statismo-shaperegistration/raw/master/data/VSD002_femur.vtk","./VSD002_femur.vtk",method = "w")
download.file(url="https://github.com/marcelluethi/statismo-shaperegistration/raw/master/data/VSD001-lm.csv","./VSD001-lm.csv",method = "w")
download.file(url="https://github.com/marcelluethi/statismo-shaperegistration/raw/master/data/VSD002-lm.csv","./VSD002-lm.csv",method = "w")

ref <- read.vtk("VSD001_femur.vtk")
tar <- read.vtk("VSD002_femur.vtk")
ref.lm <- as.matrix(read.csv("VSD001-lm.csv",row.names=1))
tar.lm <- as.matrix(read.csv("VSD002-lm.csv",row.names=1))
mymod <- statismoModelFromRepresenter(ref,kernel=list(c(50,50)),ncomp = 100,isoScale = 0.1)
mymodC <- statismoConstrainModel(mymod,tar.lm,ref.lm,2)
fit <- modelFitting(mymodC,tarmesh,iterations = 15)
}

