% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displacementField.r
\name{plot.DisplacementField}
\alias{plot.DisplacementField}
\title{visualize a displacement field}
\usage{
\method{plot}{DisplacementField}(x, lwd = 1, color = TRUE, plot = TRUE,
  ...)
}
\arguments{
\item{x}{displacement field of class "DisplacementField", e.g. created using \code{\link{createDisplacementField}}}

\item{lwd}{width of the displacement vectors}

\item{color}{logical: if TRUE, displacement vectors will be colored according to a heatmap.}

\item{plot}{if FALSE only an object of class "DisplacementPlot" is returned without graphical output}

\item{...}{additional arguments. Currently not used.}
}
\value{
invisible object of class "DisplacementPlot" that can be rendered using \code{plot}
}
\description{
visualize a displacement field
}
\seealso{
\code{\link{interpolateDisplacementField}, \link{applyDisplacementField}, \link{smoothDisplacementField}}
}

