% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displacementField.r
\name{plot.DisplacementPlot}
\alias{plot.DisplacementPlot}
\title{plot the output of plot.DisplacementField}
\usage{
\method{plot}{DisplacementPlot}(x, lwd = 1, color = TRUE,
  size = NULL, ...)
}
\arguments{
\item{x}{displacement field of class "DisplacementField", e.g. created using \code{\link{createDisplacementField}}}

\item{lwd}{width of the displacement vectors}

\item{color}{logical: if TRUE, displacement vectors will be colored according to a heatmap.}

\item{size}{size of grid points}

\item{...}{additional arguments. Currently not used.}
}
\description{
plot the output of plot.DisplacementField
}
\examples{
require(Rvcg)
data(dummyhead)
humoff <- meshOffset(dummyhead.mesh,offset=5)
dispfield <- createDisplacementField(dummyhead.mesh,humoff)
displot <- plot(dispfield,plot=FALSE)
plot(displot)
}
\seealso{
\code{\link{interpolateDisplacementField}, \link{applyDisplacementField}, \link{smoothDisplacementField}}
}
