% Generated by roxygen2 (4.0.1): do not edit by hand
\name{predictPCAcond}
\alias{predictPCAcond}
\alias{predictPCAcond.matrix}
\alias{predictPCAcond.mesh3d}
\alias{predictpPCA}
\alias{predictpPCA.matrix}
\alias{predictpPCA.mesh3d}
\alias{predictpPCA.numeric}
\title{predict or restrict a mesh or matrix based on a statistical model}
\usage{
predictPCAcond(x, model, refmesh, origSpace = TRUE, pPCA = FALSE, ...)

\method{predictPCAcond}{matrix}(x, model, refmesh = FALSE, origSpace = TRUE,
  pPCA = FALSE, ...)

\method{predictPCAcond}{mesh3d}(x, model, refmesh = FALSE, sdmax,
  origSpace = TRUE, pPCA = FALSE, ...)

predictpPCA(x, model, refmesh = FALSE, ...)

\method{predictpPCA}{matrix}(x, model, refmesh = FALSE, origSpace = TRUE,
  use.lm = NULL, sdmax, mahaprob = c("none", "chisq", "dist"), ...)

\method{predictpPCA}{mesh3d}(x, model, refmesh = FALSE, origSpace = TRUE,
  use.lm = NULL, sdmax, mahaprob = c("none", "chisq", "dist"), ...)

\method{predictpPCA}{numeric}(x, model, refmesh = FALSE, ...)
}
\arguments{
\item{x}{a matrix, a mesh3d or a vector (for pPCA models) containing standardized variables within the PC-space}

\item{model}{model of class pPCA or pPCAcond}

\item{refmesh}{if TRUE and the model contains a representer mesh, a surface mesh will be returned, coordinate matrix otherwise.}

\item{origSpace}{logical: rotate the estimation back into the original coordinate system.}

\item{pPCA}{logical: if TRUE, a constrained pPCA model is returned.
"chisq" uses the Chi-Square distribution of the squared Mahalanobisdistance, while "dist" restricts the values to
be within a multi-dimensional sphere of radius \code{sdmax}. If FALSE the probability will be determined per PC separately.}

\item{use.lm}{optional: integer vector specifying row indices of the coordinates to use for rigid registration on the model's meanshape.}

\item{sdmax}{maximum allowed standard deviation (per Principal axis) within the model space. Defines the probabilistic boundaries.}

\item{mahaprob}{character: if != "none", use mahalanobis-distance to determine overall probability (of the shape projected into the model space.}
}
\value{
\code{predictpPCA} returns a matrix/mesh3d restricted to the boundaries given by the modelspace.

\code{predictPCAcond} returns a list with
\item{estim}{matrix/mesh3d representing the mean of the restricted space}

\item{pPCA}{if \code{pPCA = TRUE} a pPCA model representing the gaussian subspace given the constraints is returned}
\item{rot}{the transformation of x into the modelspace that can be reverted by calling \code{rotreverse} from the package Morpho}
}
\description{
predict or restrict a mesh or matrix based on a statistical model
}

