% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norafilter.r
\name{read.ano.json}
\alias{read.ano.json}
\title{read landmarks from json file exported by NORA software}
\usage{
read.ano.json(x)
}
\arguments{
\item{x}{filename}
}
\description{
read landmarks from json file exported by NORA software
}
