% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeIntersect.r
\name{removeIntersect}
\alias{removeIntersect}
\title{This functions aims to translate two meshes to remove an intersection}
\usage{
removeIntersect(
  reference,
  target,
  stepsize = 0.2,
  maxit = 100,
  tol = 1,
  outside = TRUE,
  gradthresh = -Inf,
  gradn = Inf,
  realign = FALSE,
  minalign = 5,
  alignthresh = 0.5,
  visualize = FALSE,
  silent = FALSE,
  threads = 0,
  alpha = 1
)
}
\arguments{
\item{reference}{matrix or mesh}

\item{target}{mesh of class mesh3d}

\item{stepsize}{regularize the actual displacement}

\item{maxit}{maximum iterations}

\item{tol}{positive number: stop if intersection is less than \code{tol}}

\item{outside}{logical: if TRUE the reference will be placed outside of the target. Inside otherwise}

\item{gradthresh}{numerical: (abs) gradient threshold to determine convergence}

\item{gradn}{integer: number of steps to include in computing gradient}

\item{realign}{logical: if TRUE, \code{reference} will be aligned by the points closest outside/inside of the target surface.}

\item{alignthresh}{threshold which points to consider for realignment. For \code{alignthresh=.5}, all coordinates within .5 mm outside/inside of the targtet will be used to realign the shape.}

\item{visualize}{logical: if TRUE watch the approximation}

\item{silent}{logical: suppress messages}

\item{threads}{number of threads to use for closest point search (0=auto)}

\item{alpha}{value with \code{0 <= alpha <=1} to render surface opacity of the target mesh.}
}
\description{
This functions aims to translate two meshes to remove an intersection. This is done based on signed distance
}
\details{
This is a quite simple attempt to remove an intersection after the alignment of two meshes. This is achieved by translating \code{reference} along the average difference vectors of all vertices
}
