% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectIsolated.r
\name{selectIsolated}
\alias{selectIsolated}
\title{interactively select isolated pieces from a mesh}
\usage{
selectIsolated(mesh, maxpiece = 10)
}
\arguments{
\item{mesh}{triangular mesh of class mesh3d}

\item{maxpiece}{integer: the n-largest (number of vertices) pieces to consider}
}
\value{
returns a mesh with all selected parts merged and colored
}
\description{
interactively select isolated pieces from a mesh
}

