% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorMesh.r
\name{setVertColor}
\alias{setVertColor}
\title{set the per-vertex colors of an object of class mesh3d}
\usage{
setVertColor(x, colorvec)
}
\arguments{
\item{x}{mesh3d}

\item{colorvec}{vector of lenght \code{ncol(x$vb)} containing HEX rgb values for vertex colors.}
}
\value{
mesh with updated vertex colors
}
\description{
set the per-vertex colors of an object of class mesh3d
}
