% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displacementField.r
\name{smoothDisplacementField}
\alias{smoothDisplacementField}
\title{smooth a displacement field using Gaussian smoothing}
\usage{
smoothDisplacementField(dispfield, k = 10, sigma = 20, type = c("Gauss",
  "Laplace", "Exponential"), threads = 1)
}
\arguments{
\item{dispfield}{displacement field created using \code{\link{createDisplacementField}}}

\item{k}{integer: number of k closest points to evaluate.}

\item{sigma}{sigma controls the weight of the neighbourhood by defining the standard-deviation for the gaussian smoothing}

\item{type}{kernel function for smoothing are "Gauss","Laplace" and "Exponential"}

\item{threads}{integer: number of threads to use for computing the interpolation.}
}
\description{
smooth a displacement field using Gaussian smoothing
}
\seealso{
\code{\link{interpolateDisplacementField}, \link{applyDisplacementField}, \link{plotDisplacementField}}
}

