% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posteriorDeform.r
\name{subsampleDeform}
\alias{subsampleDeform}
\title{Deforms a reference to a target based on a TPS or AmbergDeform}
\usage{
subsampleDeform(
  reference,
  target,
  partsample = NULL,
  samplenum = 1000,
  distance = 1e+10,
  slide = 3,
  bending = TRUE,
  ray = FALSE,
  Amberg = FALSE,
  rhotol = pi/2,
  reflm = NULL,
  tarlm = NULL,
  forceLM = FALSE,
  silent = FALSE,
  threads = 1,
  ...
)
}
\arguments{
\item{reference}{reference mesh}

\item{target}{target mesh}

\item{partsample}{predetermined sample points on the reference mesh}

\item{samplenum}{integer: if partsample=NULL, this specifies the number of coordinates sampled on the model mean}

\item{distance}{numeric: constrain maximum distance to mark target point as appropriate}

\item{slide}{integer: if > 0 the valid correspondences on the model instance will be relaxed minimizing bending energy/procrustes distance.}

\item{bending}{logical: if TRUE, the coordinates on the model instance are relaxed using bending energy, Procrustes distance otherwise}

\item{ray}{logical: if TRUE, the closest point search will be performed along the normals only}

\item{Amberg}{if TRUE the deformation will use the function \code{\link{AmbergDeformSpam}} and \code{\link{tps3d}} otherwise}

\item{rhotol}{maximal tolerated angle between normals to be considered a valid match}

\item{reflm}{matrix containing 3D landmarks on the reference}

\item{tarlm}{matrix containing 3D landmarks on the target surface}

\item{forceLM}{if TRUE, predfined landmarks \code{modlm} are not allowed to slide. For cases with high uncertainty, this can lead to unwanted mesh distortions.}

\item{silent}{logical: supress debug output}

\item{threads}{integer: number of threads to use for tps interpolation (set to 1 if using openblas, or otherwise it can become instable)}

\item{...}{additional parameters passed to \code{\link{AmbergDeformSpam}}.}
}
\value{
returns a deformed version of a model instance fitted to the target
}
\description{
Deforms a (pre-aligned) reference to a target based on a TPS/AmbergDeform and automatically sampled sliding semi-landmarks
}
\note{
Please note that it is required to align the target mesh to the reference mesh beforehand. This can be performed using the function \code{\link{icp}}, for example.
}
