% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bary2point.r
\name{transferPoints}
\alias{transferPoints}
\title{transfer points between two registered meshes}
\usage{
transferPoints(x, mesh1, mesh2, tolwarn = 0.01)
}
\arguments{
\item{x}{a matrix with 3D-coordinates (or a mesh), positioned on \code{mesh1}}

\item{mesh1}{a mesh on which \code{x} is placed}

\item{mesh2}{a mesh with vertices and faces corresponding to  \code{mesh1} (e.g. registered using \code{\link{gaussMatch}}).}

\item{tolwarn}{numeric: if at least one coordinate of \code{x} is further away from \code{mesh1} than \code{tolwarn}, a warning will be issued.}
}
\value{
returns a matrix containing \code{x} tranfered to \code{mesh2}.
}
\description{
transfer points between two registered meshes
}
\details{
the function gets the barycentric coordinates of  \code{x} on \code{mesh1} and uses them to find the corresponding positions on \code{mesh2}
}
\examples{
require(Rvcg)
data(humface)
#extract 300 random points from humface
coords <- vcgSample(humface,200,"pd")

#move original mesh
transface <- translate3d(humface, 10, 10 ,10)
##extract coordinates
newcoord <- transferPoints(coords, humface, transface)
\dontrun{
require(rgl)
wire3d(transface, col=3)
spheres3d(newcoord)
}
}

