% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/trivolume.r
\name{trivolume}
\alias{trivolume}
\title{Calculate volume between two states of a mesh}
\usage{
trivolume(mesh1, mesh2)
}
\arguments{
\item{mesh1}{triangular mesh in first state}

\item{mesh2}{triangular mesh in second state}
}
\value{
Volume
}
\description{
Calculate volume between two states of a mesh by dividing it into prisms
}

