% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateMeshlist.r
\name{updateVertices}
\alias{updateVertices}
\title{replace vertices of a mesh}
\usage{
updateVertices(x, matrix)
}
\arguments{
\item{x}{mesh of class mesh3d}

\item{matrix}{with nrow(matrix) == amount of vertices in \code{x}}
}
\value{
returns updated mesh
}
\description{
replace vertices of a mesh
}

