% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norafilter.r
\name{write.ano.json}
\alias{write.ano.json}
\title{write landmarks to json file readable by NORA software}
\usage{
write.ano.json(x, filename = dataname, labels = dataname, size = 1)
}
\arguments{
\item{x}{\code{k x 3} matrix containing landmarks}

\item{filename}{character: output file name}

\item{labels}{character vector of length \code{k} containing landmark names}

\item{size}{integer: voxel size around landmark}
}
\description{
write landmarks to json file readable by NORA software
}
