% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displacementField.r
\name{applyDisplacementField}
\alias{applyDisplacementField}
\title{apply a discrete displacement field to a set of points/mesh in its domain}
\usage{
applyDisplacementField(dispfield, points, k = 10, sigma = 20,
  type = c("Gauss", "Laplace", "Exponential"), gamma = 1, threads = 1)
}
\arguments{
\item{dispfield}{displacement field of class \code{DisplacementField}}

\item{points}{matrix or mesh3d at which to evaluate the interpolated displacement field}

\item{k}{integer: number of k closest points to evaluate.}

\item{sigma}{sigma controls the weight of the neighbourhood by defining the standard-deviation for the gaussian smoothing}

\item{type}{kernel function for smoothing are "Gauss","Laplace" and "Exponential"}

\item{gamma}{dampening factor (displacement vectors will be divided by \code{gamma}}

\item{threads}{integer: number of threads to use for computing the interpolation.}
}
\value{
returns the displaced version of points
}
\description{
apply a discrete displacement field to a set of points/mesh in its domain by applying the gaussian smoothed interpolation based of k closest neighbours
}
\note{
if points is identical to the domain of the displacement field, no interpolation will be performed.
}

