% Generated by roxygen2 (4.0.0): do not edit by hand
\name{buildAffineAmberg}
\alias{buildAffineAmberg}
\title{subfunction of AmbergDeformSpam}
\usage{
buildAffineAmberg(mesh)
}
\arguments{
\item{mesh}{triangular mesh}
}
\description{
subroutine preparing vertex assortment in order to create S using an 'arcface' matrix
}

