% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Amberg2011.r
\name{createArcNode}
\alias{createArcNode}
\title{subfunction of AmbergDeformSpam}
\usage{
createArcNode(mesh)
}
\arguments{
\item{mesh}{triangular mesh}
}
\description{
create per face arcnode matrix to differentiate per-face affine trafo based on per edge approximation
}

