% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insideBBox.r
\name{getMeshBox}
\alias{getMeshBox}
\title{compute the bounding box of a triangular mesh}
\usage{
getMeshBox(mesh, tri = FALSE, extend = 0, pca = TRUE)
}
\arguments{
\item{mesh}{triangular mesh of class "mesh3d"}

\item{tri}{logical: if TRUE, triangular mesh representing the bounding box is returned}

\item{extend}{numeric: offset extending the bbox (or shrinking if negative)}

\item{pca}{if TRUE, the box will be aligned by the principal axes.}
}
\value{
returns a matrix with the corners of the box or a triangular mesh if tri=TRUE.
}
\description{
compute the bounding box of a triangular mesh
}
\examples{
require(Rvcg)
data(humface)
bbox <- getMeshBox(humface)
\dontrun{
require(rgl)
spheres3d(bbox)
}
bbmesh <- getMeshBox(humface,tri=TRUE)
\dontrun{
shade3d(bbmesh,alpha=0.4)
wire3d(humface)
}
}

