% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/insideBBox.r
\name{makeBox}
\alias{makeBox}
\title{create a bounding box as mesh}
\usage{
makeBox(corners)
}
\arguments{
\item{corners}{a 8 x 3 matrix defining the corners of a bounding box. With first 4 rows containing the corners of one square and 5:8 their opposing counterparts.}
}
\description{
create a bounding box as mesh
}

