\name{mesheR-package}
\Rdversion{1.1}
\alias{mesheR-package}
\alias{mesheR}
\docType{package}
\title{
Meshing tools

}
\description{
A toolbox for meshing operations: elastic mesh mapping, Iteratively
closest point matching, vertex selection, ...,
}
\details{
\tabular{ll}{
Package: mesheR
Type: Package
Version: 0.3
Date: 2013-06-28
License: GPL
LazyLoad: yes
}

}
\author{Stefan Schlager


Maintainer: Stefan Schlager <zarquon42@gmail.com>

}
\references{
To be announced
}

\keyword{ package }


