\name{mesheR-package}
\Rdversion{1.1}
\alias{mesheR-package}
\alias{mesheR}
\docType{package}
\title{
Meshing tools

}
\description{
A toolbox for meshing operations: elastic mesh mapping, Iteratively
closest point matching, vertex selection, ...,
}
\details{
\tabular{ll}{
Package: \tab mesheR\cr
Type: \tab Package\cr
Version: \tab 0.2.1\cr
Date: \tab 2013-03-18\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}

}
\author{Stefan Schlager


Maintainer: Stefan Schlager <zarquon42@gmail.com>

}
\references{
To be announced
}

\keyword{ package }


