% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesheR-package.R
\docType{package}
\name{mesheR-package}
\alias{mesheR-package}
\alias{mesheR}
\title{Meshing tools}
\description{
A toolbox for meshing operations: elastic mesh mapping, Iteratively closest point matching, vertex selection, ...,
}
\details{
\tabular{ll}{
Package: \tab mesheR\cr
Type: \tab Package\cr
Version: \tab 0.4.200211\cr
Date: \tab 2020-02-11\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr }
}
\references{
To be announced
}
\author{
Stefan Schlager

Maintainer: Stefan Schlager <zarquon42@gmail.com>
}
\keyword{package}
