\docType{package}
\name{mesheR-package}
\alias{mesheR}
\alias{mesheR-package}
\title{Meshing tools}
\description{
A toolbox for meshing operations: elastic mesh mapping,
Iteratively closest point matching, vertex selection, ...,
}
\details{
\tabular{ll}{ Package: \tab mesheR\cr Type: \tab Package\cr
Version: \tab 0.4.140506\cr Date: \tab 2014-05-06\cr
License: \tab GPL\cr LazyLoad: \tab yes\cr }
}
\author{
Stefan Schlager

Maintainer: Stefan Schlager <zarquon42@gmail.com>
}
\references{
To be announced
}
\keyword{package}

