% Generated by roxygen2 (4.0.2): do not edit by hand
\name{normcheck}
\alias{normcheck}
\title{compare normal directions between two states of a mesh}
\usage{
normcheck(mesh1, mesh2)
}
\arguments{
\item{mesh1}{triangular mesh}

\item{mesh2}{triangular mesh}

\item{colwise}{logical: if TRUE, the columns will be compared.}
}
\value{
numeric vector containing angles between corresponding normals
}
\description{
compare normal directions between two states of a mesh
}

