% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icp.r
\name{normcheck}
\alias{normcheck}
\title{compare normal directions between two states of a mesh}
\usage{
normcheck(mesh1, mesh2, threads = 0)
}
\arguments{
\item{mesh1}{triangular mesh}

\item{mesh2}{triangular mesh}

\item{threads}{number of threads to use}
}
\value{
numeric vector containing angles between corresponding normals
}
\description{
compare normal directions between two states of a mesh
}
