% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insideBBox.r
\name{outsideBBox}
\alias{outsideBBox}
\title{get vertex indices of vertices outside a bounding box}
\usage{
outsideBBox(x, corners, outside = TRUE)
}
\arguments{
\item{x}{a k x 3 matrix or a triangular mesh of class "mesh3d"}

\item{corners}{a 8 x 3 matrix defining the corners of a bounding box, or a triangular mesh of class "mesh3d" representing the meshBox}

\item{outside}{logical: if TRUE the indices outside the box are returned, those inside}
}
\value{
returns the indices of coordinates inside/outside the bounding box.
}
\description{
get vertex indices of vertices outside a bounding box
}
\examples{
require(Rvcg)
require(Morpho)

data(humface)
bbox <- getMeshBox(humface,extend=-0.2)
bboxmesh <-  getMeshBox(humface,extend=-0.2,tri=TRUE)
\dontrun{
require(rgl)
spheres3d(bbox)
}
outside <- outsideBBox(humface,bbox)
humshrink <- rmVertex(humface,outside)
\dontrun{
wire3d(humshrink,col=3)#inside
wire3d(humface,col=2)#outside
shade3d(bboxmesh,alpha=0.4)
}
}

