% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displacementField.r
\name{plotDisplacementField}
\alias{plotDisplacementField}
\title{visualize a displacement field}
\usage{
plotDisplacementField(dispfield, lwd = 1, colored = TRUE)
}
\arguments{
\item{dispfield}{displacement field of class \code{displacement_field}}

\item{lwd}{width of the displacement vectors}

\item{colored}{logical: if TRUE, displacement vectors will be colored according to a heatmap.}
}
\description{
visualize a displacement field
}

