% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignnointersect.r
\name{removeIntersect}
\alias{removeIntersect}
\title{This functions aims to translate two meshes to remove an intersection}
\usage{
removeIntersect(reference, target, stepsize = 0.2, maxit = 100, tol = 1,
  outside = TRUE, rot = FALSE)
}
\arguments{
\item{reference}{matrix or mesh}

\item{target}{mesh of class mesh3d}

\item{stepsize}{regularize the actual displacement}

\item{maxit}{maximum iterations}

\item{tol}{positive number: stop if intersection is less than \code{tol}}

\item{outside}{logical: if TRUE the reference will be placed outside of the target. Inside otherwise}
}
\description{
This functions aims to translate two meshes to remove an intersection. This is done based on signed distance
}
\details{
This is a quite simple attempt to remove an intersection after the alignment of two meshes. This is achieved by translating \code{reference} along the average difference vectors of all vertices
}
