% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rmInternals}
\alias{rmInternals}
\title{remove vertices within a convex surface mesh}
\usage{
rmInternals(mesh, mindist = NULL, explode = FALSE, center = NULL)
}
\arguments{
\item{mesh}{triangular mesh of class "mesh3d"}

\item{mindist}{numeric: restrict to those vertices whose intersection point
is further away as mindist.}

\item{explode}{logical: instead of surface normals use the direction outward
from a center of the mesh. Per default, the center of the meshes bounding
box is used.}

\item{center}{numeric vector of length 3, used in the case "explode=TRUE",
if another center is required.}
}
\value{
returns a cleaned mesh
}
\description{
remove vertices within a convex surface mesh by checking if they intersect
the mesh along their normals.
}
\details{
This only works for convex meshes, for concave substructures this might
result in unwanted removals.
}
\author{
Stefan Schlager
}
\seealso{
\code{\link{rmVertex}}
}
\keyword{~kwd1}
\keyword{~kwd2}

