% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displacementField.r
\name{smoothDisplacementfield}
\alias{smoothDisplacementfield}
\title{smooth a displacement field using Gaussian smoothing}
\usage{
smoothDisplacementfield(dispfield, k = 10, sigma = 20, threads = 1)
}
\arguments{
\item{dispfield}{displacement field created using \code{\link{createDisplacementField}}}

\item{k}{integer: number of k closest points to evaluate.}

\item{sigma}{sigma controls the weight of the neighbourhood by defining the standard-deviation for the gaussian smoothing}

\item{threads}{integer: number of threads to use for computing the interpolation.}
}
\description{
smooth a displacement field using Gaussian smoothing
}

