% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displacementField.r
\name{smoothDisplacementField}
\alias{smoothDisplacementField}
\title{smooth a displacement field using Gaussian smoothing}
\usage{
smoothDisplacementField(dispfield, k = 10, sigma = 20,
  type = c("Gauss", "Laplace", "Exponential", "Bspline", "TPS"),
  iterations = 1, subsample = 2000, lambda = 1e-08, threads = 0)
}
\arguments{
\item{dispfield}{displacement field of class "DisplacementField", e.g. created using \code{\link{createDisplacementField}}}

\item{k}{integer: number of k closest points to evaluate.}

\item{sigma}{kernel bandwidth used for smoothing. For all kernels except B-spline, sigma controls the importance of the neighbourhood by defining the bandwidth of the smoothing kernel. For B-spline it defines the support (the higher, the "wobblier" the deformation field can become.}

\item{type}{kernel function for smoothing are "Gauss","Laplace", "Exponential", "Bspline" and "TPS" (or any abbreviation thereof).}

\item{iterations}{number of iterations to run}

\item{subsample}{integer: amount to subsample the field in case of type="TPS"}

\item{lambda}{smoothing factor for TPS}

\item{threads}{integer: number of threads to use for computing the interpolation.}
}
\description{
smooth a displacement field using Gaussian smoothing
}
\seealso{
\code{\link{interpolateDisplacementField}, \link{applyDisplacementField}, \link{plot.DisplacementField}}
}
