% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warpmovieMulti.R
\name{warpmovieMulti}
\alias{warpmovieMulti}
\title{make a warpmovie with multiple registered meshes}
\usage{
warpmovieMulti(..., n, col = NULL, folder = NULL, movie = "warpmovie",
  add = FALSE, close = TRUE, countbegin = 0, ask = TRUE,
  whichcolor = NULL, align = TRUE, scale = FALSE, mixcolor = TRUE,
  shade = c("s", "w", "b"))
}
\arguments{
\item{\dots}{registered meshes or a list containing registered meshes.}

\item{n}{amount of intermediate images between two meshes}

\item{col}{mesh color - overrides options \code{whichcolor} and \code{mixcolor}}

\item{folder}{character: output folder for created images (optional)}

\item{movie}{character: name of the output files}

\item{add}{logical: if TRUE, the movie will be added to the focussed rgl-windows.}

\item{close}{logical: if TRUE, the rgl window will be closed when finished. width and 200 the height of the image.}

\item{countbegin}{integer: number to start image sequence.}

\item{ask}{logical: if TRUE, the viewpoint can be selected manually.}

\item{whichcolor}{select which mesh's vertexcolors to use.}

\item{align}{logical: align all meshes to the first one}

\item{scale}{if \code{align=TRUE} this controls if scaling is allowed.}

\item{mixcolor}{logical: if available existing colors are subsequently mixed using \code{\link{mixColorMesh}}.}

\item{shade}{select shade: "s"=shade3d, "w"= wire3d, "b"=both}
}
\description{
make a warpmovie with multiple registered meshes - this is a pimped version of warpmovie3d from package 'Morpho'
}
\examples{
require(Morpho)
data(nose)
redmesh <- shortnose.mesh
redmesh$material$color <- matrix("#FF0000",dim(shortnose.mesh$it))
bluemesh <- shortnose.mesh
bluemesh$material$color <- matrix("#0000FF",dim(shortnose.mesh$it))
\dontrun{
warpmovieMulti(bluemesh, redmesh, n=15)
}
}
