% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiremesh3d.r
\name{wiremesh3d}
\alias{wiremesh3d}
\alias{wiremesh3d.mesh3d}
\title{create a shaded wiremesh}
\usage{
wiremesh3d(x, col = "white", specular = 1, ...)

\method{wiremesh3d}{mesh3d}(x, col = "white", specular = 1, ...)
}
\arguments{
\item{x}{mesh3d}

\item{col}{color of shaded mesh}

\item{specular}{parameter from rgl}

\item{...}{additional parameters passed to shade3d from package rgl.}
}
\value{
a vector containing the shape ids
}
\description{
create a shaded wiremesh
}

